+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/EngineSession.kt
@@ -1152,4 +1152,11 @@ abstract class EngineSession(
      * @param enabled True if the activity is in picture-in-picture mode.
      */
     open fun onPipModeChanged(enabled: Boolean) = Unit
+
+    /**
+     * Asynchronously determine if the page loaded in this session uses a QWAC.
+     *
+     * @param onResult Callback to call with the QWAC or null if none.
+     */
+    open fun qwacStatus(onResult: (X509Certificate?) -> Unit) = Unit
 }